group: expand group lpid and ch4 gpid into 64bit

Approved-by: Yanfei Guo <yguo@anl.gov>

+++ src/mpid/ch4/src/ch4_init.c
@@ -774,7 +774,7 @@ int MPID_Free_mem(void *user_buf)
     goto fn_exit;
 }
 
-int MPID_Comm_get_lpid(MPIR_Comm * comm_ptr, int idx, int *lpid_ptr, bool is_remote)
+int MPID_Comm_get_lpid(MPIR_Comm * comm_ptr, int idx, uint64_t * lpid_ptr, bool is_remote)
 {
     int mpi_errno = MPI_SUCCESS;
     int avtid = 0, lpid = 0;
